implemented the basic reusable crud repository and also user specific crud repository

+++ app/repositories/base.py
@@ -1,3 +1,4 @@
+from datetime import datetime, timezone
 from typing import TypeVar, Generic, Optional
 from uuid import UUID
 
@@ -13,22 +14,28 @@ class BaseCrud(Generic[ModelT]):
         self.model = model
         self.session = session
 
-    async def _get_by_id(self, uid: UUID) -> Optional[ModelT]:
+    async def get_by_id(self, uid: UUID) -> Optional[ModelT]:
         return await self.session.get(self.model, uid)
 
-    async def _save(self, entity: SQLModel):
+    async def save(self, entity: ModelT) -> ModelT:
         self.session.add(entity)
         await self.session.commit()
         await self.session.refresh(entity)
         return entity
 
-    async def _update(self, entity: SQLModel):
+    async def update(self, entity: ModelT) -> ModelT:
         self.session.add(entity)
         await self.session.commit()
         return entity
 
-    async def _delete(self, entity: SQLModel):
-        await self.session.delete(entity)
+    async def delete(self, entity: ModelT, soft: bool = True) -> None:
+        if soft and hasattr(entity, "is_deleted"):
+            entity.is_deleted = True
+            entity.deleted_at = datetime.now(tz=timezone.utc)
+            entity.is_active = False
+        else:
+            await self.session.delete(entity)
+
         await self.session.commit()
